+++ chat/templates/chat_index.html
@@ -1,0 +1,36 @@
+<!doctype html>
+<html lang="en">
+<head>
+    <meta charset="UTF-8">
+    <meta name="viewport" content="width=device-width, initial-scale=1.0">
+    <title>聊天功能</title>
+</head>
+<body>
+    <h1>聊天功能</h1>
+    <form id="chat-form">
+        <textarea name="message" rows="4" cols="50" placeholder="请输入您的消息"></textarea><br>
+        <input type="submit" value="发送">
+    </form>
+    <div id="chat-history"></div>
+
+    <script>
+        document.getElementById('chat-form').addEventListener('submit', function(event) {
+            event.preventDefault();
+            const formData = new FormData(this);
+
+            fetch('/chat/chat', {
+                method: 'POST',
+                body: formData
+            })
+            .then(response => response.json())
+            .then(data => {
+                const chatHistoryDiv = document.getElementById('chat-history');
+                chatHistoryDiv.innerHTML += `<p>Assistant: ${data.response}</p>`;
+            })
+            .catch(error => {
+                console.error('Error:', error);
+            });
+        });
+    </script>
+</body>
+</html>
